requirements.txt
PyFunctional == 1.1.2
PyMuPDF      == 1.14.3

boxsdk       == 2.0.0
pyzotero     == 1.3.8

requests     == 2.20.0

# Google Sheets
google-api-python-client == 1.7.8
google-auth              == 1.6.3
google-auth-httplib2     == 0.0.3
google-auth-oauthlib     == 0.3.0
